Renaming of TokenService
… to TokenOperations
