+++ src/main/java/org/trello4j/Trello.java
@@ -8,12 +8,13 @@ import org.trello4j.core.ListOperations;
 import org.trello4j.core.MemberOperations;
 import org.trello4j.core.NotificationOperations;
 import org.trello4j.core.OrganizationOperations;
+import org.trello4j.core.TokenOperations;
 import org.trello4j.model.Type;
 
 /**
  * The Interface Trello.
  */
-public interface Trello extends OrganizationOperations, NotificationOperations, BoardOperations, CardOperations, ListOperations, MemberOperations, ChecklistOperations, TokenService, ActionOperations {
+public interface Trello extends OrganizationOperations, NotificationOperations, BoardOperations, CardOperations, ListOperations, MemberOperations, ChecklistOperations, TokenOperations, ActionOperations {
 
 	/**
 	 * Gets the type.
